hw3/pyproject.toml
[project]
name = "hw3-q-learning"
version = "0.1.0"
description = "CS285 HW3: Q-Learning and Actor-Critic Algorithms"
readme = "README.md"
requires-python = "==3.10.*"
dependencies = [
    "matplotlib>=3.10.8",
    "ml-collections>=1.1.0",
    "modal>=1.3.0.post1",
    "opencv-python<4.12.0.88",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "wandb[media]>=0.23.1",
    "gym==0.25.2",
    "numpy<2.0.0",
    "pygame>=2.5.2",
    "pyyaml>=6.0",
    "gradescope-utils>=0.5.0",
    "glfw>=2.10.0",
    "mujoco>=3.8.0",
]

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
